disable after click
Former-commit-id: 14c45d0b01398aae764d4513e2a2214448e728e1

+++ RavenQuestionnaire.Core/RavenQuestionnaire.Core/Services/CompleteQuestionnaireUploaderService.cs
@@ -110,7 +110,8 @@ namespace RavenQuestionnaire.Core.Services
         public void DeleteCompleteQuestionnaire(string id)
         {
             var entity = _questionRepository.Load(id);
-            this._questionRepository.Remove(entity);
+            if(entity!=null)
+                this._questionRepository.Remove(entity);
             var statEntity = _statisticsRepository.Load(IdUtil.CreateStatisticId(IdUtil.ParseId(id)));
             if (statEntity != null)
                 this._statisticsRepository.Remove(statEntity);
